+++ lib/Perl/Tidy/FileWriter.pm
@@ -16,7 +16,7 @@
 package Perl::Tidy::FileWriter;
 use strict;
 use warnings;
-our $VERSION = '20260204';
+our $VERSION = '20260204.01';
 use Carp;
 
 use constant DEVEL_MODE   => 0;