+++ lib/Perl/Tidy/IndentationItem.pm
@@ -9,7 +9,7 @@ package Perl::Tidy::IndentationItem;
 use strict;
 use warnings;
 
-our $VERSION = '20260204';
+our $VERSION = '20260204.01';
 
 BEGIN {